TBA: Live Computation now faster and setted to productComputer2
git-svn-id: svn+ssh://erebor.esa.informatik.tu-darmstadt.de/home/wimi/svn/GaalopCompiler/gaalop-2.0/branches/tbaAndGapp@1023 4235ee25-d3ef-403b-b6ed-692c26853d4b

+++ tba/src/main/java/de/gaalop/productComputer2/GeoProductCalculator.java
@@ -11,7 +11,7 @@ import java.util.List;
  */
 public class GeoProductCalculator implements ProductCalculator {
 
-    public SumOfBlades calculate(Blade blade1, Blade blade2, HashMap<Integer, Integer> baseSquares) {
+    public SumOfBlades calculate(Blade blade1, Blade blade2, HashMap<Integer, Byte> baseSquares) {
         //approach method from the dissertation: Daniel Fontijne, "Efficient Implementation of Geometric Algebra" (2007)
 
         LinkedList<Integer> list1 = new LinkedList(blade1.getBaseVectors());
